Split tests and create new samples
