+++ examples/react-native/__tests__/ts/sync/get-subscriptions.test.tsx
@@ -1,0 +1,101 @@
+// :snippet-start: get-subscriptions
+import React, {useEffect} from 'react';
+import {AppProvider, UserProvider, createRealmContext} from '@realm/react';
+// import object model
+import Turtle from '../Models/Turtle';
+import Bird from '../Models/Bird';
+import {Text, FlatList} from 'react-native';
+// :remove-start:
+import {App, Credentials} from 'realm';
+import {useApp, Realm} from '@realm/react';
+import {render, waitFor} from '@testing-library/react-native';
+
+const APP_ID = 'js-flexible-oseso';
+
+let numSubs: number;
+
+function LogIn() {
+  const app = useApp();
+
+  useEffect(() => {
+    app.logIn(Credentials.anonymous());
+  }, []);
+
+  return <></>;
+}
+// :remove-end:
+
+const config = {
+  // Pass in imported object models
+  schema: [Turtle, Bird],
+};
+
+const RealmContext = createRealmContext(config);
+const {RealmProvider} = RealmContext;
+
+function AppWrapper() {
+  return (
+    <AppProvider id={APP_ID}>
+      <UserProvider fallback={LogIn}>
+        <RealmProvider
+          sync={{
+            flexible: true,
+            // Set initial subscription for all objects of type `Turtle`.
+            // You can add more than one subscription here.
+            initialSubscriptions: {
+              update(subs, realm) {
+                subs.add(realm.objects(Turtle));
+              },
+            },
+            onError: console.log,
+          }}>
+          <SubscriptionManager />
+        </RealmProvider>
+      </UserProvider>
+    </AppProvider>
+  );
+}
+
+function SubscriptionManager() {
+  const {useRealm, useQuery} = RealmContext;
+  const realm = useRealm();
+
+  // Returns a subscription set that contains all subscriptions.
+  const allSubscriptions = realm.subscriptions;
+
+  // Pass object model to useQuery and filter results.
+  // This does not create a subscription.
+  const seenBirds = useQuery(Bird).filtered('haveSeen == true');
+
+  useEffect(() => {
+    realm.subscriptions.update(mutableSubs => {
+      // Create subscription for filtered results.
+      mutableSubs.add(seenBirds);
+    });
+    numSubs = realm.subscriptions.length; // :remove:
+  });
+
+  return (
+    <FlatList
+      data={allSubscriptions}
+      keyExtractor={subscription => subscription.id.toString()}
+      renderItem={({item}) => <Text>{item.name}</Text>}
+    />
+  );
+}
+// :snippet-end:
+
+afterEach(async () => {
+  await App.getApp(APP_ID).currentUser?.logOut();
+  Realm.deleteFile(config);
+});
+
+test('Instantiate AppWrapper and children correctly', async () => {
+  render(<AppWrapper />);
+  await waitFor(
+    () => {
+      expect(numSubs).toBe(2);
+    },
+    {timeout: 2000},
+  );
+});
